ref/pyrcicr_ref/requirements.txt
#asttokens==2.4.1
#cffi==1.16.0
#contourpy==1.2.0
#cycler==0.12.1
#decorator==5.1.1
#executing==2.0.1
#fonttools==4.50.0
#iniconfig==2.0.0
#ipython==8.22.2
#jedi==0.19.1
#Jinja2==3.1.3
#kiwisolver==1.4.5
#MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
#opencv-python==4.9.0.80
#packaging==24.0
pandas==2.2.1
#parso==0.8.3
#pexpect==4.9.0
pillow==10.2.0
#pluggy==1.4.0
#prompt-toolkit==3.0.43
#ptyprocess==0.7.0
#pure-eval==0.2.2
#pycparser==2.21
#Pygments==2.17.2
#pyparsing==3.1.2
#pytest==8.1.1
#python-dateutil==2.9.0.post0
#pytz==2024.1
scipy==1.12.0
#six==1.16.0
#stack-data==0.6.3
tqdm==4.66.2
#traitlets==5.14.2
#tzdata==2024.1
#tzlocal==5.2
#wcwidth==0.2.13
